requirements.txt
cython
annoy==1.16.3
Flask==1.1.2
Flask-Limiter==1.3.1
gunicorn[gevent]==20.0.4
ImageHash==4.1.0
piexif==1.1.3
Pillow
PyYAML
tweepy==3.8.0
Werkzeug==1.0.1
beautifulsoup4
tldextract
scikit-learn==0.22.2.post1
opencv-python
opencv-contrib-python
joblib==0.15.1
requests-cache==0.5.2
redis==3.5.2
youtube_dl
bsddb3==6.2.7
cachetools==4.2.2
